Updated to enable selection of I2C BUS ID to support other I2C controllers e.g. CP2112
The original code was configured to use the BUS ID of the onboard Pi4 controller (bus_id = 1) that is used for DIY builds of PiKVM - and I suspect also for the purchased v3 and v4 variants.

In order to enable these v3 / v4 variants to be used (without physical modification) with the ATX Switch it is necessary to provide an external I2C controller capability that can use existing interfaces on the v3 / v4 (at least on the v4+). The CP2112 debug board enables a USB port on the PiKVM to be used to provide the I2C controls required.

However the BUS ID allocated by the controller differs from the internal interface, and could potentially be different for different controllers. This required the CP2112 BUS ID to be determined and then set in the ATX Switch scripts. This only required changing two (2) hard coded values so was minimal effort, but I thought it better to modify the code to enable the I2C BUS ID (i2c_bus_id) to be specified in the atx_switch_initialisation.yaml file